requirements.txt
psutil
pyserial
tabulate